+++ .env.example
@@ -1,2 +1,4 @@
 API_URL=
 BASE_URL=
+APP_STORE_URL=
+PLAY_STORE_URL=
