+++ src/features/force-update-modal/force-update-modal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Modal, Text, View } from 'react-native';
 
-import { useValidateAppVersion } from 'hooks/force-update';
+import { useValidateAppVersion } from 'hooks/useValidateAppVersion';
 
 import { translate } from 'localization/hooks';
 
@@ -10,10 +10,10 @@ import { styles } from './force-update-modal.styles';
 interface ForceUpdateProps {}
 
 const ForceUpdateModal = ({}: ForceUpdateProps) => {
-  const { showModal, openStore } = useValidateAppVersion();
+  const { isUpdateAvailable, openStore } = useValidateAppVersion();
 
   return (
-    <Modal animationType="slide" transparent={true} visible={showModal}>
+    <Modal animationType="slide" transparent={true} visible={isUpdateAvailable}>
       <View style={styles.centeredView}>
         <View style={styles.modalView}>
           <Text style={styles.modalText}>{translate('alert.update')}</Text>
